удоление карточки в movies
